Update to use HTTPS Flickr API URLs
Copy Paste of the Flickr email:

We see that you own an active API key for the Flickr API, and wanted make sure you know about some upcoming changes to the API related to SSL. 

As you may have read on the Flickr code blog, we've updated all of the API endpoints to fully support SSL. You no longer need to use a special subdomain to access the API securely, and we want to make sure that all API communication adopts SSL. 

Many of you have moved over to using SSL already, which is great! If you haven't, it shouldn't be much more work than updating your apps to call the API at: https://api.flickr.com/ 

We'll be updating the blog with information as we get close to the switchover date, but here's the most important info: 

New API keys will be issued for HTTPS access only: as of 6 May 2014
First blackout test window: 3 June 2014, 10:00-12:00 Pacific Daylight Time (PDT)
Second blackout test window: 17 June 2014, 18:00-20:00 (PDT)
Non-SSL API deprecated: 27 June 2014, 10:00 (PDT) 

During the blackout tests and after the 27th of June, calls to the Flickr API that are not over SSL will fail with an HTTP error code 403. We will also return an error message with (Flickr) code 95 and msg "SSL is required". 

Here are example error responses: 

XML (REST): 

<?xml version="1.0" encoding="utf-8" ?> <rsp stat="fail"> <err code="95" msg="SSL is required" /> </rsp> 

JSONP: 

jsonFlickrApi({ "stat": "fail", "code": "95", "message": "SSL is required" }) 

JSON: 

{ "stat": "fail", "code": "95", "message": "SSL is required" } 

We realize that this change might be more difficult for some. We will follow the Developer Support Group closely, so please let us hear your questions. 

If you have a question about the API transition to SSL that you feel is unique to your situation or is more appropriate to handle via conversation directly with Flickr, send an email to flickr-api-help@yahoo-inc.com. We will monitor that mailbox through July 2014. 

Thank you for being a member of our API community,
The Flickreenos